+++ updater/site/WoWInterface.py
